Added Camera movment + made map bigger

+++ Vector.pde
@@ -1,12 +1,12 @@
 class Vector {
-  float x, y;
+  int x, y;
   
   Vector () {
    this.x = 0;
    this.y = 0;
   }
   
-  Vector (float x, float y) {
+  Vector (int x, int y) {
     this.x = x;
     this.y = y;
   }
